feat: refactor type imports and consolidate type definitions for badges and languages
