+++ src/shared/types/themes.type.ts
@@ -1,0 +1,18 @@
+export interface ThemeOverrides {
+    theme?: string;
+    bgColor?: string;
+    borderColor?: string;
+    textColor?: string;
+    titleColor?: string;
+}
+
+export interface Theme {
+    titleColor: string;
+    textColor: string;
+    iconColor: string;
+    bgColor: string;
+    borderColor: string;
+    fontName?: string;
+    fontFamily?: string;
+    fontUrl?: string;
+}
