+++ src/shared/utils/themes/graph.ts
@@ -1,4 +1,4 @@
-import { Theme } from "../../../types.js";
+import { Theme } from "../../types/themes.type.js";
 
 /**
  * Graph-optimized themes for the /graph activity heatmap.
